requirements-dev.txt
Flask==1.0.2 
line-bot-sdk==1.15.0
google-api-python-client==1.8.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.41
python-dotenv==0.13.0